+++ bindings/vue/CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 dev
 ----
 * VOnsNavigator: Added `popPage` optional prop.
+* VOnsTabbar: Fix update issue.
+* VOnsCarousel: Fix update issue.
 
 2.0.0-alpha.1
 ----