+++ bindings/vue/CHANGELOG.md
@@ -3,6 +3,13 @@ CHANGELOG
 
 dev
 ----
+* Autostyling: Fix for 'material' modifier.
+* VOnsNavigator: Added `popPage` optional prop.
+* VOnsTabbar: Fix update issue.
+* VOnsCarousel: Fix update issue.
+
+2.0.0-alpha.1
+----
 
 ### New Features